// tests/setup.d.ts
export declare const testUserData: {
    email: string;
    password: string;
    firstName: string;
    lastName: string;
    phone: string;
    farmName: string;
    farmAddress: string;
    farmCity: string;
    farmProvince: string;
};
export declare const testLoginData: {
    email: string;
    password: string;
};
//# sourceMappingURL=setup.d.ts.map